+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - Update RSS feed URL, the feed is built from the weekly site https://weekly.bestofjs.org/rss/trends.xml
+- Add commit hook, using [pretty-quick](https://github.com/azz/pretty-quick)
 
 ## 0.12.0 - 2018-06-01